Fix the build with Clang 11
* Xsynchronize returns a pointer, so use jlong rather than jint in the
  wrapper prototype

Authored by: dim@FreeBSD.org

+++ src/java.desktop/unix/classes/sun/awt/X11/XlibWrapper.java
@@ -529,7 +529,7 @@ static native String XSetLocaleModifiers(String modifier_list);
           display   Specifies the connection to the X server.
           onoff     Specifies a Boolean value that indicates whether to enable or disable synchronization.
      */
-    static native int XSynchronize(long display, boolean onoff);
+    static native long XSynchronize(long display, boolean onoff);
 
     /**
      * Extracts an X event that can be processed in a secondary loop.
